feat(plot): Modify input of plot to model
