+++ pystoned/__init__.py
@@ -8,7 +8,7 @@ from . import CQERG
 from . import dataset
 from . import DEA
 from . import FDH
-from . import Plot
+from . import plot
 from . import ICNLS
 from . import ICQER
 from . import StoNED
@@ -24,7 +24,7 @@ __all__ = [
     'dataset',
     'DEA',
     'FDH',
-    'Plot',
+    'plot',
     'ICNLS',
     'ICQER',
     'StoNED',
